+++ deploy/complete/terraform/mushop-utilities.tf
@@ -3,11 +3,11 @@
 # 
 
 # Create namespace mushop-utilities for supporting services
-resource "kubernetes_namespace" "mushop_utilities_namespace" {
+resource "kubernetes_namespace" "cluster_utilities_namespace" {
   metadata {
     name = "mushop-utilities"
   }
-  depends_on = [oci_containerengine_node_pool.oke_mushop_node_pool]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 
 # MuShop Utilities helm charts
@@ -18,8 +18,8 @@ resource "helm_release" "prometheus" {
   name       = "prometheus"
   repository = local.helm_repository.prometheus
   chart      = "prometheus"
-  version    = "13.5.0"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  version    = "13.6.0"
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = false
 
   values = [
@@ -37,18 +37,47 @@ resource "helm_release" "grafana" {
   name       = "mushop-utils-grafana" # mushop-utils included to be backwards compatible to the docs and setup chart install
   repository = local.helm_repository.grafana
   chart      = "grafana"
-  version    = "6.4.8"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  version    = "6.7.1"
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = false
 
   set {
     name  = "grafana\\.ini.server.root_url"
     value = "%(protocol)s://%(domain)s:%(http_port)s/grafana"
-    type  = "string"
   }
 
   values = [
-    file("${path.module}/chart-values/grafana-values.yaml"),
+    file("${path.module}/chart-values/grafana-values.yaml"), <<EOF
+datasources: 
+  datasources.yaml:
+    apiVersion: 1
+    datasources:
+    - name: Prometheus
+      type: prometheus
+      url: http://prometheus-server.${kubernetes_namespace.cluster_utilities_namespace.id}.svc.cluster.local
+      access: proxy
+      isDefault: true
+      disableDeletion: true
+      editable: false
+    - name: Oracle Cloud Infrastructure Metrics
+      type: oci-metrics-datasource
+      access: proxy
+      disableDeletion: true
+      editable: true
+      jsonData:
+        tenancyOCID: ${var.tenancy_ocid}
+        defaultRegion: ${var.region}
+        environment: "OCI Instance"
+    - name: Oracle Cloud Infrastructure Logs
+      type: oci-logs-datasource
+      access: proxy
+      disableDeletion: true
+      editable: true
+      jsonData:
+        tenancyOCID: ${var.tenancy_ocid}
+        defaultRegion: ${var.region}
+        environment: "OCI Instance"
+EOF
   ]
 
   depends_on = [helm_release.ingress_nginx] # Ugly workaround because of the oci pvc provisioner not be able to wait for the node be active and retry.
@@ -62,7 +91,7 @@ resource "helm_release" "metrics_server" {
   repository = local.helm_repository.stable
   chart      = "metrics-server"
   version    = "2.11.4"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = false
 
   values = [
@@ -80,8 +109,8 @@ resource "helm_release" "ingress_nginx" {
   name       = "mushop-utils-ingress-nginx" # mushop-utils included to be backwards compatible to the docs and setup chart install
   repository = local.helm_repository.ingress_nginx
   chart      = "ingress-nginx"
-  version    = "3.23.0"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  version    = "3.26.0"
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = true
 
   set {
@@ -92,7 +121,6 @@ resource "helm_release" "ingress_nginx" {
   set {
     name  = "controller.service.annotations.service\\.beta\\.kubernetes\\.io/oci-load-balancer-shape"
     value = var.ingress_load_balancer_shape
-    type  = "string"
   }
 
   timeout = 1800 # workaround to wait the node be active for other charts
@@ -105,7 +133,7 @@ resource "helm_release" "svc-cat" {
   repository = local.helm_repository.svc_catalog
   chart      = "catalog"
   version    = "0.3.1"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = false
 
   depends_on = [helm_release.ingress_nginx] # Ugly workaround because of the oci pvc provisioner not be able to wait for the node be active and retry.
@@ -120,7 +148,7 @@ resource "helm_release" "cert_manager" {
   repository = local.helm_repository.jetstack
   chart      = "cert-manager"
   version    = "1.2.0"
-  namespace  = kubernetes_namespace.mushop_utilities_namespace.id
+  namespace  = kubernetes_namespace.cluster_utilities_namespace.id
   wait       = false
 
   set {
@@ -131,4 +159,35 @@ resource "helm_release" "cert_manager" {
   depends_on = [helm_release.ingress_nginx] # Ugly workaround because of the oci pvc provisioner not be able to wait for the node be active and retry.
 
   count = var.cert_manager_enabled ? 1 : 0
+}
+
+# MuShop Datasources for outputs
+## Kubernetes Service: mushop-utils-ingress-nginx-controller
+data "kubernetes_service" "mushop_ingress" {
+  metadata {
+    name      = "mushop-utils-ingress-nginx-controller" # mushop-utils name included to be backwards compatible to the docs and setup chart install
+    namespace = kubernetes_namespace.cluster_utilities_namespace.id
+  }
+  depends_on = [helm_release.ingress_nginx]
+}
+
+## Kubernetes Secret: Grafana Admin Password
+data "kubernetes_secret" "mushop_utils_grafana" {
+  metadata {
+    name      = "mushop-utils-grafana"
+    namespace = kubernetes_namespace.cluster_utilities_namespace.id
+  }
+  depends_on = [helm_release.grafana, helm_release.mushop]
+}
+
+locals {
+  # Helm repos
+  helm_repository = {
+    stable        = "https://charts.helm.sh/stable"
+    ingress_nginx = "https://kubernetes.github.io/ingress-nginx"
+    jetstack      = "https://charts.jetstack.io"                        # cert-manager
+    svc_catalog   = "https://kubernetes-sigs.github.io/service-catalog" # Service Catalog
+    grafana       = "https://grafana.github.io/helm-charts"
+    prometheus    = "https://prometheus-community.github.io/helm-charts"
+  }
 }