+++ deploy/complete/terraform/oke-variables.tf
@@ -1,0 +1,185 @@
+# Copyright (c) 2021 Oracle and/or its affiliates. All rights reserved.
+# Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
+# 
+
+# OKE Variables
+## OKE Cluster Details
+variable "app_name" {
+  default = "MuShop"
+}
+variable "cluster_name" {
+  default     = "MuShop-cluster"
+  description = "OKE cluster name prefix"
+}
+variable "create_new_oke_cluster" {
+  default     = true
+  description = "Creates a new OKE cluster, node pool and network resources"
+}
+variable "existent_oke_cluster_id" {
+  default     = ""
+  description = "Cluster Id of the existent OKE"
+}
+variable "create_new_compartment_for_oke" {
+  default     = false
+  description = "Creates new compartment for OKE Nodes and OCI Services deployed.  NOTE: The creation of the compartment increases the deployment time by at least 3 minutes, and can increate by 15 minutes when destroying"
+}
+variable "oke_compartment_description" {
+  default = "Compartment for OKE, Nodes and Services"
+}
+variable "cluster_options_add_ons_is_kubernetes_dashboard_enabled" {
+  default = false
+}
+variable "cluster_options_admission_controller_options_is_pod_security_policy_enabled" {
+  description = "If true: The pod security policy admission controller will use pod security policies to restrict the pods accepted into the cluster."
+  default     = false
+}
+
+## OKE Visibility (Workers and Endpoint)
+
+variable "cluster_workers_visibility" {
+  default     = "Private"
+  description = "The Kubernetes worker nodes that are created will be hosted in public or private subnet(s)"
+
+  validation {
+    condition     = var.cluster_workers_visibility == "Private" || var.cluster_workers_visibility == "Public"
+    error_message = "Sorry, but cluster visibility can only be Private or Public."
+  }
+}
+
+variable "cluster_endpoint_visibility" {
+  default     = "Public"
+  description = "The Kubernetes cluster that is created will be hosted on a public subnet with a public IP address auto-assigned or on a private subnet. If Private, additional configuration will be necessary to run kubectl commands"
+
+  validation {
+    condition     = var.cluster_endpoint_visibility == "Private" || var.cluster_endpoint_visibility == "Public"
+    error_message = "Sorry, but cluster endpoint visibility can only be Private or Public."
+  }
+}
+
+## OKE Encryption details
+variable "use_encryption_from_oci_vault" {
+  default     = false
+  description = "By default, Oracle manages the keys that encrypts Kubernetes Secrets at Rest in Etcd, but you can choose a key from a vault that you have access to, if you want greater control over the key's lifecycle and how it's used"
+}
+variable "create_new_encryption_key" {
+  default     = false
+  description = "Creates new vault and key on OCI Vault/Key Management/KMS and assign to boot volume of the worker nodes"
+}
+variable "existent_encryption_key_id" {
+  default     = ""
+  description = "Use an existent master encryption key to encrypt boot volume and object storage bucket. NOTE: If the key resides in a different compartment or in a different tenancy, make sure you have the proper policies to access, or the provision of the worker nodes will fail"
+}
+variable "create_vault_policies_for_group" {
+  default     = false
+  description = "Creates policies to allow the user applying the stack to manage vault and keys. If you are on the Administrators group or already have the policies for a compartment, this policy is not needed. If you do not have access to allow the policy, ask your administrator to include it for you"
+}
+variable "user_admin_group_for_vault_policy" {
+  default     = "Administrators"
+  description = "User Identity Group to allow manage vault and keys. The user running the Terraform scripts or Applying the ORM Stack need to be on this group"
+}
+
+## OKE Node Pool Details
+variable "node_pool_name" {
+  default     = "pool1"
+  description = "Name of the node pool"
+}
+variable "k8s_version" {
+  default     = "Latest"
+  description = "Kubernetes version installed on your master and worker nodes"
+}
+variable "num_pool_workers" {
+  default     = 3
+  description = "The number of worker nodes in the node pool"
+}
+variable "node_pool_shape" {
+  default     = "VM.Standard.E3.Flex"
+  description = "A shape is a template that determines the number of CPUs, amount of memory, and other resources allocated to a newly created instance for the Worker Node"
+}
+variable "node_pool_node_shape_config_ocpus" {
+  default     = "1" # Only used if flex shape is selected
+  description = "You can customize the number of OCPUs to a flexible shape"
+}
+variable "node_pool_node_shape_config_memory_in_gbs" {
+  default     = "16" # Only used if flex shape is selected
+  description = "You can customize the amount of memory allocated to a flexible shape"
+}
+variable "node_pool_boot_volume_size_in_gbs" {
+  default     = "60"
+  description = "Specify a custom boot volume size (in GB)"
+}
+variable "image_operating_system" {
+  default     = "Oracle Linux"
+  description = "The OS/image installed on all nodes in the node pool."
+}
+variable "image_operating_system_version" {
+  default     = "7.9"
+  description = "The OS/image version installed on all nodes in the node pool."
+}
+variable "generate_public_ssh_key" {
+  default = true
+}
+variable "public_ssh_key" {
+  default     = ""
+  description = "In order to access your private nodes with a public SSH key you will need to set up a bastion host (a.k.a. jump box). If using public nodes, bastion is not needed. Left blank to not import keys."
+}
+
+# Network Details
+## CIDRs
+variable "network_cidrs" {
+  type = map(string)
+
+  default = {
+    VCN-CIDR                      = "10.20.0.0/16"
+    SUBNET-REGIONAL-CIDR          = "10.20.10.0/24"
+    LB-SUBNET-REGIONAL-CIDR       = "10.20.20.0/24"
+    ENDPOINT-SUBNET-REGIONAL-CIDR = "10.20.0.0/28"
+    ALL-CIDR                      = "0.0.0.0/0"
+    PODS-CIDR                     = "10.244.0.0/16"
+    KUBERNETES-SERVICE-CIDR       = "10.96.0.0/16"
+  }
+}
+
+# OCI Provider
+variable "tenancy_ocid" {}
+variable "compartment_ocid" {}
+variable "region" {}
+variable "user_ocid" {
+  default = ""
+}
+variable "fingerprint" {
+  default = ""
+}
+variable "private_key_path" {
+  default = ""
+}
+
+# Create Dynamic Group and Policies
+variable "create_dynamic_group_for_nodes_in_compartment" {
+  default     = true
+  description = "Creates dynamic group of Nodes in the compartment. Note: You need to have proper rights on the Tenancy. If you only have rights in a compartment, uncheck and ask you administrator to create the Dynamic Group for you"
+}
+variable "existent_dynamic_group_for_nodes_in_compartment" {
+  default     = ""
+  description = "Enter previous created Dynamic Group for the policies"
+}
+variable "create_compartment_policies" {
+  default     = true
+  description = "Creates policies that will reside on the compartment. e.g.: Policies to support Cluster Autoscaler, OCI Logging datasource on Grafana"
+}
+variable "create_tenancy_policies" {
+  default     = true
+  description = "Creates policies that need to reside on the tenancy. e.g.: Policies to support OCI Metrics datasource on Grafana"
+}
+
+# ORM Schema visual control variables
+variable "show_advanced" {
+  default = false
+}
+
+# Dictionary Locals
+locals {
+  compute_flexible_shapes = [
+    "VM.Standard.E3.Flex",
+    "VM.Standard.E4.Flex"
+  ]
+}