+++ deploy/complete/terraform/oke.tf
@@ -1,15 +1,19 @@
-# Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2020, 2021 Oracle and/or its affiliates. All rights reserved.
 # Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
 # 
 
-resource "oci_containerengine_cluster" "oke_mushop_cluster" {
-  compartment_id     = var.compartment_ocid
-  kubernetes_version = var.k8s_version
+resource "oci_containerengine_cluster" "oke_cluster" {
+  compartment_id     = local.oke_compartment_ocid
+  kubernetes_version = var.k8s_version == "Latest" ? local.cluster_k8s_latest_version : var.k8s_version
   name               = "${var.cluster_name}-${random_string.deploy_id.result}"
-  vcn_id             = oci_core_virtual_network.oke_mushop_vcn[0].id
+  vcn_id             = oci_core_virtual_network.oke_vcn[0].id
 
+  endpoint_config {
+    is_public_ip_enabled = (var.cluster_endpoint_visibility == "Private") ? false : true
+    subnet_id            = oci_core_subnet.oke_k8s_endpoint_subnet[0].id
+  }
   options {
-    service_lb_subnet_ids = [oci_core_subnet.oke_mushop_lb_subnet[0].id]
+    service_lb_subnet_ids = [oci_core_subnet.oke_lb_subnet[0].id]
     add_ons {
       is_kubernetes_dashboard_enabled = var.cluster_options_add_ons_is_kubernetes_dashboard_enabled
       is_tiller_enabled               = false # Default is false, left here for reference
@@ -22,15 +26,15 @@ resource "oci_containerengine_cluster" "oke_mushop_cluster" {
       pods_cidr     = lookup(var.network_cidrs, "PODS-CIDR")
     }
   }
-  kms_key_id = var.use_encryption ? var.encryption_key_id : null
+  kms_key_id = var.use_encryption_from_oci_vault ? (var.create_new_encryption_key ? oci_kms_key.mushop_key[0].id : var.existent_encryption_key_id) : null
 
   count = var.create_new_oke_cluster ? 1 : 0
 }
 
-resource "oci_containerengine_node_pool" "oke_mushop_node_pool" {
-  cluster_id         = oci_containerengine_cluster.oke_mushop_cluster[0].id
-  compartment_id     = var.compartment_ocid
-  kubernetes_version = var.k8s_version
+resource "oci_containerengine_node_pool" "oke_node_pool" {
+  cluster_id         = oci_containerengine_cluster.oke_cluster[0].id
+  compartment_id     = local.oke_compartment_ocid
+  kubernetes_version = var.k8s_version == "Latest" ? local.node_pool_k8s_latest_version : var.k8s_version
   name               = var.node_pool_name
   node_shape         = var.node_pool_shape
   ssh_public_key     = var.generate_public_ssh_key ? tls_private_key.oke_worker_node_ssh_key.public_key_openssh : var.public_ssh_key
@@ -41,7 +45,7 @@ resource "oci_containerengine_node_pool" "oke_mushop_node_pool" {
 
       content {
         availability_domain = placement_configs.value.name
-        subnet_id           = oci_core_subnet.oke_mushop_subnet[0].id
+        subnet_id           = oci_core_subnet.oke_nodes_subnet[0].id
       }
     }
     size = var.num_pool_workers
@@ -56,8 +60,9 @@ resource "oci_containerengine_node_pool" "oke_mushop_node_pool" {
   }
 
   node_source_details {
-    source_type = "IMAGE"
-    image_id    = lookup(data.oci_core_images.node_pool_images.images[0], "id")
+    source_type             = "IMAGE"
+    image_id                = lookup(data.oci_core_images.node_pool_images.images[0], "id")
+    boot_volume_size_in_gbs = var.node_pool_boot_volume_size_in_gbs
   }
 
   initial_node_labels {
@@ -66,6 +71,18 @@ resource "oci_containerengine_node_pool" "oke_mushop_node_pool" {
   }
 
   count = var.create_new_oke_cluster ? 1 : 0
+}
+
+resource "oci_identity_compartment" "oke_compartment" {
+  compartment_id = var.compartment_ocid
+  name           = "${trimspace(var.app_name)}-${random_string.deploy_id.result}"
+  description    = "${var.app_name} ${var.oke_compartment_description} (Deployment ${random_string.deploy_id.result})"
+  enable_delete  = true
+
+  count = var.create_new_compartment_for_oke ? 1 : 0
+}
+locals {
+  oke_compartment_ocid = var.create_new_compartment_for_oke ? oci_identity_compartment.oke_compartment.0.id : var.compartment_ocid
 }
 
 # Local kubeconfig for when using Terraform locally. Not used by Oracle Resource Manager
@@ -78,6 +95,12 @@ resource "local_file" "kubeconfig" {
 resource "tls_private_key" "oke_worker_node_ssh_key" {
   algorithm = "RSA"
   rsa_bits  = 2048
+}
+
+# Get OKE options
+locals {
+  cluster_k8s_latest_version   = reverse(sort(data.oci_containerengine_cluster_option.oke.kubernetes_versions))[0]
+  node_pool_k8s_latest_version = reverse(sort(data.oci_containerengine_node_pool_option.oke.kubernetes_versions))[0]
 }
 
 # Checks if is using Flexible Compute Shapes